Separate DependentField into its own component
