+++ components/klaudsolcms/modals/modal_body/CollectionTypeBody.js
@@ -2,34 +2,7 @@ import { useState, useContext, useEffect } from "react";
 import { Formik, Form, Field, useFormikContext, useField } from "formik";
 import { loadEntityTypes } from "@/components/reducers/actions";
 import RootContext from "@/components/contexts/RootContext";
-
-// TO BE SEPARATED INTO ANOTHER COMPONENT
-function DependentField(props) {
-  const { values, setFieldValue } = useFormikContext();
-  const [field] = useField(props);
-  const { format, inputToMirror } = props;
-  const value = values[inputToMirror];
-
-  // Remove inputToMirror & format pros because it is not a legit input
-  // attribute and it will give us a warning of we put it on input
-  const filteredProps = Object.keys(props).reduce((acc, curr) => {
-    if (curr === "inputToMirror" || curr === "format") return acc;
-
-    acc[curr] = props[curr];
-    return acc;
-  }, {});
-
-  useEffect(() => {
-    const outputVal = format(value);
-    setFieldValue(props.name, outputVal);
-  }, [value]);
-
-  return <input {...filteredProps} {...field} />;
-}
-
-DependentField.defaultProps = {
-  format: (value) => value,
-};
+import DependentField from "@/components/klaudsolcms/modals/modal_body/DependentField";
 
 export default function CollectionTypeBody({ formRef }) {
   const { state: rootState, dispatch: rootDispatch } = useContext(RootContext);
